Implement SO_RERROR
Summary:
SO_RERROR indicates that receive buffer overflows should be handled as errors.
Historically receive buffer overflows have been ignored and programs
could not tell if they missed messages or messages had been truncated
because of overflows.  Since programs historically do not expect to get
receive overflow errors, this behavior is not the default.

This is really **really** important for programs that use route(4) to keep in sync
with the system. If we loose a message then we need to reload the full system
state, otherwise the behaviour from that point is undefined and can lead
to chasing bogus bug reports.

Test Plan:
Download and compile this reproducer: https://xenity.marples.name/~roy/overflow.c

If it cannot detect overflow then your system suffers from it.

Reviewers: #network, #manpages, #transport, melifaro

Reviewed By: #network, melifaro

Subscribers: philip, cy, pi, gnn, afedorov, bcr, tuexen, emaste, imp, ae, bz, donner, melifaro

Differential Revision: https://reviews.freebsd.org/D26652